requirements.txt
Django==1.4
Markdown==2.1.1
MySQL-python==1.2.3
BeautifulSoup==3.2.0
cheetah==2.4.4
modwsgideploy==0.4.20dev
python-memcached==1.48
html5lib==0.95
slimit==0.7.4
PIL==1.1.7
South==0.7.5
anyjson==0.3.1
celery==2.5.3
django-annoying==0.7.6
django-autoslug==1.5.0
django-compressor==1.1.2
django-extensions==0.8
django-profiles==0.2
django-registration==0.8
django-mptt==0.5.2
clint==0.3.1
